+++ src/petro/box/avcc.hpp
@@ -31,7 +31,7 @@ public:
 public:
     avcc(std::weak_ptr<box> parent);
 
-    void read(uint64_t size, byte_stream& bs);
+    void read(uint32_t size, byte_stream& bs);
 
     virtual std::string describe() const;
