+++ src/petro/box/bxml.cpp
@@ -17,7 +17,7 @@ bxml::bxml(std::weak_ptr<box> parent) :
     box(bxml::TYPE, parent)
 { }
 
-void bxml::read(uint64_t size, byte_stream& bs)
+void bxml::read(uint32_t size, byte_stream& bs)
 {
     box::read(size, bs);
     bs.skip(m_remaining_bytes);
