+++ src/petro/box/bxml.hpp
@@ -26,7 +26,7 @@ public:
 public:
     bxml(std::weak_ptr<box> parent);
 
-    void read(uint64_t size, byte_stream& bs);
+    void read(uint32_t size, byte_stream& bs);
 };
 }
 }
