+++ src/petro/box/co64.hpp
@@ -28,7 +28,7 @@ public:
         full_box(co64::TYPE, parent)
     { }
 
-    void read(uint64_t size, byte_stream& bs)
+    void read(uint32_t size, byte_stream& bs)
     {
         full_box::read(size, bs);
         m_entry_count = bs.read_uint32_t();
