+++ src/petro/box/esds.hpp
@@ -30,7 +30,7 @@ public:
 
 public:
     esds(std::weak_ptr<box> parent);
-    void read(uint64_t size, byte_stream& bs);
+    void read(uint32_t size, byte_stream& bs);
     std::string describe() const;
 
     descriptor_type descriptor() const;
