+++ src/petro/box/mfhd.hpp
@@ -28,7 +28,7 @@ public:
         box(mfhd::TYPE, parent)
     { }
 
-    void read(uint64_t size, byte_stream& bs)
+    void read(uint32_t size, byte_stream& bs)
     {
         box::read(size, bs);
         bs.skip(m_remaining_bytes);
