+++ src/petro/box/root.hpp
@@ -26,7 +26,7 @@ public:
     { }
 
     /// hide box implementation
-    void read(uint64_t size, byte_stream& bs) override
+    void read(uint32_t size, byte_stream& bs) override
     {
         (void) size;
         (void) bs;
