+++ src/petro/box/stts.hpp
@@ -41,7 +41,7 @@ public:
         full_box(stts::TYPE, parent)
     { }
 
-    void read(uint64_t size, byte_stream& bs)
+    void read(uint32_t size, byte_stream& bs)
     {
         full_box::read(size, bs);
         m_entry_count = bs.read_uint32_t();
