+++ src/petro/box/stz2.hpp
@@ -29,7 +29,7 @@ public:
         full_box(stz2::TYPE, parent)
     { }
 
-    void read(uint64_t size, byte_stream& bs)
+    void read(uint32_t size, byte_stream& bs)
     {
         full_box::read(size, bs);
         bs.skip(3);
